Rework URLs for JWT instead of Base-64, change fetchers
Fetchers: pass options instead of a URL. Decoding a URL is a messy affair.
For files and ActiveRecords it already turned out that we would need to do
escaping and de-escaping to make the URL valid. While the basic idea - dispatching
to a specific Fetcher based on the scheme or name (like `file`, `http` etc.) is sane,
the actual way it worked was pretty bleak. Replace this with the fetcher name
and fetcher params.

URLs: use a single JWT for access to the image. While the URL becomes longer, we have
a number of clear benefits:

* JWTs can be generated by something else than ImageVise
* The Base64 encoding is always URL-safe and does not require any special manipulation
* The signature is applied and encoded in a standard manner